+++ src/mympd_api/mympd_api_handler.c
@@ -5,6 +5,7 @@
 */
 
 #include "compile_time.h"
+#include "mpd/playlist.h"
 #include "mympd_api_handler.h"
 
 #include "../lib/album_cache.h"
@@ -1316,9 +1317,24 @@ void mympd_api_handler(struct t_partition_state *partition_state, struct t_work_
             }
             break;
         case MYMPD_API_QUEUE_SAVE:
-            if (json_get_string(request->data, "$.params.plist", 1, FILENAME_LEN_MAX, &sds_buf1, vcb_isfilename, &error) == true) {
-                rc = mpd_run_save(partition_state->conn, sds_buf1);
-                response->data = mympd_respond_with_error_or_ok(partition_state, response->data, request->cmd_id, request->id, rc, "mpd_run_save", &result);
+            if (json_get_string(request->data, "$.params.plist", 1, FILENAME_LEN_MAX, &sds_buf1, vcb_isfilename, &error) == true &&
+                json_get_string(request->data, "$.params.mode", 1, NAME_LEN_MAX, &sds_buf2, vcb_isalnum, &error) == true)
+            {
+                if (mympd_state->mpd_state->feat_advqueue == true) {
+                    enum mpd_queue_save_mode save_mode = mpd_parse_queue_save_mode(sds_buf2);
+                    if (save_mode != MPD_QUEUE_SAVE_MODE_UNKNOWN) {
+                        rc = mpd_run_save_queue(partition_state->conn, sds_buf1, save_mode);
+                        response->data = mympd_respond_with_error_or_ok(partition_state, response->data, request->cmd_id, request->id, rc, "mpd_run_save_queue", &result);
+                    }
+                    else {
+                        response->data = jsonrpc_respond_message(response->data, request->cmd_id, request->id,
+                            JSONRPC_FACILITY_QUEUE, JSONRPC_SEVERITY_ERROR, "Unknown queue save mode");
+                    }
+                }
+                else {
+                    rc = mpd_run_save(partition_state->conn, sds_buf1);
+                    response->data = mympd_respond_with_error_or_ok(partition_state, response->data, request->cmd_id, request->id, rc, "mpd_run_save", &result);
+                }
             }
             break;
         case MYMPD_API_QUEUE_SEARCH: {