+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This minor release adds support of new MPD 0.24 features.
 - Feat: support consume oneshot (MPD 0.24) #837
 - Feat: autoconfiguration for pcre support (MPD 0.24) #843
 - Feat: support starts_with filter expression (MPD 0.24) #843
+- Feat: support queue save modes (MPD 0.24) #848
 - Upd: libmympdclient 1.0.15
 
 ***